// sillage-web/src/types/fragrance.ts
export type Fragrance = {
    id: string
    userId: string
    name: string
    house: string
    imageUrl: string
    concentration: number
    genre: number
    isManual: boolean
    generalNotes: string
    mainAccords: string
    seasonRanking: string
    occasionRanking: string
    notes: string
    dateAdded: string
}